ChameliaV2/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "h-jepa"
version = "0.1.0"
description = "Hierarchical Joint-Embedding Predictive Architecture (H-JEPA) implementation in PyTorch"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "H-JEPA Team"}
]
keywords = ["deep-learning", "self-supervised-learning", "computer-vision", "jepa", "vision-transformers"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "timm>=0.9.0",
    "einops>=0.7.0",
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "PyYAML>=6.0",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
    "gymnasium>=0.29.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.9.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
cognitive = [
    "faiss-cpu>=1.13.2",
    "h5py>=3.16.0",
    "loralib>=0.1.2",
    "lancedb>=0.30.2",
    "pyarrow>=23.0.1",
    "stitch-core>=0.1.29",
    "botorch>=0.17.2",
    "gpytorch>=1.15.2",
    "transformers>=5.5.0",
    "sentencepiece>=0.2.1",
]
interactive = [
    "gymnasium>=0.29.1",
    "pettingzoo>=1.24.0",
    "python-chess>=1.999",
]
protein_dti = [
    "biopython>=1.83",
    "PyTDC>=0.4.1",
    "rdkit>=2023.9.0",
]

[project.urls]
Homepage = "https://github.com/jonwiggins/H-JEPA"
Documentation = "https://github.com/jonwiggins/H-JEPA#readme"
Repository = "https://github.com/jonwiggins/H-JEPA"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "E712",   # comparison to True/False (common in tests with torch tensors)
    "B006",   # mutable default arguments (too many to refactor safely)
    "B007",   # unused loop control variable (prefix with _ is noise)
    "B008",   # do not perform function calls in argument defaults
    "B028",   # no explicit stacklevel in warnings.warn
    "B905",   # zip() without strict= parameter
    "E741",   # ambiguous variable name
    "UP038",  # isinstance(x, (A, B)) is valid and more readable than X | Y
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = false  # different stubs across platforms make this unreliable
warn_no_return = true
warn_unreachable = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
no_implicit_optional = true
disallow_any_generics = true
disallow_subclassing_any = true
ignore_missing_imports = true
no_implicit_reexport = false
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "numpy.*",
    "matplotlib.*",
    "PIL.*",
    "tqdm.*",
    "wandb.*",
    "einops.*",
    "timm.*",
    "sklearn.*",
    "seaborn.*",
    "tensorboard.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "src.models.hjepa"
disallow_untyped_defs = false
disallow_any_generics = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
